Put notification page in new router add container to layout with nav
